P7 lib aar

See merge request android-terminal/pos!6

+++ p7libapi/src/main/java/ru/petrolplus/pos/p7LibApi/dto/DebitParamsDto.kt
@@ -10,11 +10,13 @@ package ru.petrolplus.pos.p7LibApi.dto
  * @property pinBlock - Введенный PIN предъявителем карты в виде PIN-блока зашифрованного публичной частью
  * RSA ключа и Nonce, предоставленного управляющему приложению ранее на этапе DetectCard
  */
+
+//todo: при возможности, использовать val и убрать инициализацию (требует существенной переработки JNI)
 class DebitParamsDto(
     var serviceWhat: Int = 0,
     var serviceFrom: Int = 0,
-    var amount: Double = 0.0,
-    var price: Double = 0.0,
-    var sum: Double = 0.0,
-    var pinBlock: ByteArray = byteArrayOf(),
+    var amount: Long = 0L,
+    var price: Long = 0L,
+    var sum: Long = 0L,
+    var pinBlock: ByteArray = byteArrayOf()
 )
