+++ p7libapi/src/main/java/ru/petrolplus/pos/p7LibApi/dto/LibInfoDto.kt
@@ -2,15 +2,16 @@ package ru.petrolplus.pos.p7LibApi.dto
 
 /**
  * Класс который содержит информацию о номере эсквайра и номер терминала, а также номер версии библиотеки.
- * TODO -  Юрий добавь описание
- * @property acquirerId -
- * @property terminalNum -
- * @property majorVersion -
- * @property minerVersion -
+ * @property acquirerId - ID экваэра
+ * @property terminalNum - Номер терминала
+ * @property majorVersion - Старший номер версии библиотеки
+ * @property minerVersion - Младший номер версии бибилиотеки
  */
+
+//todo: при возможности, использовать val и убрать инициализацию (требует существенной переработки JNI)
 class LibInfoDto(
-    var acquirerId: Int = 0,
-    var terminalNum: Int = 0,
-    var majorVersion: Int = 0,
-    var minerVersion: Int = 0
+    var acquirerId: Long = 0,
+    var terminalNum: Long = 0,
+    var majorVersion: Long = 0,
+    var minerVersion: Long = 0
 )
